ref: Improve travis webhook logging for valid payloads
