+++ zeus/testutils/fixtures.py
@@ -1,3 +1,4 @@
+import json
 import os
 import pytest
 
@@ -299,3 +300,15 @@ def git_repo_config():
     yield RepoConfig(url, path, remote_path, commits)
 
     check_call("rm -rf %s" % (root,), shell=True)
+
+
+@pytest.fixture(scope="session")
+def sample_travis_build_commit(fixture_path):
+    with open(os.path.join(fixture_path, "travis-build-commit.json"), "rb") as fp:
+        return json.load(fp)
+
+
+@pytest.fixture(scope="session")
+def sample_travis_build_pr(fixture_path):
+    with open(os.path.join(fixture_path, "travis-build-pull-request.json"), "rb") as fp:
+        return json.load(fp)
